+++ lib/modules/test/test_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: test
-// Updated at: 2026-08-16 11:16:22
-// Commit: test(test): implement widget integration tests for test
+// Updated at: 2026-08-17 12:41:15
+// Commit: test(test): add unit tests for test business logic
 
 class testComponentState {
-  final String commitRef = '3635';
-  final String lastMessage = 'test(test): implement widget integration tests for test';
-  final DateTime timestamp = DateTime.parse('2026-08-16T11:16:22+05:00');
+  final String commitRef = '3667';
+  final String lastMessage = 'test(test): add unit tests for test business logic';
+  final DateTime timestamp = DateTime.parse('2026-08-17T12:41:15+05:00');
 }